backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
python-dotenv
google-generativeai
python-multipart
